feat: Restructuring of components, creation of a main layout for UserProfile and UserContact
