+++ src/app/(panel)/teacher/profile/contact/page.tsx
@@ -1,14 +1,6 @@
-import { HeaderSection, UserContact } from '@/modules/core'
-
+import { UserContact } from "@/modules/core"
 export default function Page() {
   return (
-    <main>
-      <HeaderSection
-        title="Perfil de usuario"
-        subtitle="Registrar, actualizar datos relevantes de mi usuario"
-        disableAddButton
-      />
       <UserContact/>
-    </main>
   )
 }
