+++ src/app/(panel)/teacher/profile/page.tsx
@@ -1,14 +1,6 @@
-import { HeaderSection, UserProfile } from '@/modules/core'
-
+import { UserProfile } from "@/modules/core"
 export default function Page() {
   return (
-    <main>
-      <HeaderSection
-        title="Perfil de usuario"
-        subtitle="Registrar, actualizar datos relevantes de mi usuario"
-        disableAddButton
-      />
       <UserProfile/>
-    </main>
   )
 }
