+++ src/modules/core/components/UserProfile/UserContact.tsx
@@ -1,29 +1,11 @@
 'use client'
-
-import Link from "next/dist/client/link"
-import { LayoutAsideSection } from "../../layouts"
 import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
-import { Button, buttonVariants } from "@/components/ui/button"
+import { Button } from "@/components/ui/button"
 import { usePathname } from "next/navigation"
 
 export const UserContact = () => {
-    const pathname = usePathname()
-
-    console.log('pathname', pathname)
-
     return (
-        <LayoutAsideSection aside={<div className="flex flex-col p-5 gap-4 w-[286px] bg-white">
-            <h1 className="font-bold">Menu</h1>
-            <Link href={"/teacher/profile"} className="hover:bg-slate-100 hover:border rounded-md text-sm">
-                Datos Personales
-            </Link>
-            <Link href={pathname} className={"hover:bg-slate-100" + buttonVariants({ variant: 'secondary' })}>
-                <span className="text-sm">
-                    Datos de Contacto
-                </span>
-            </Link>
-        </div>}>
             <div className="flex flex-col gap-6 w-full">
                 <form className="flex flex-col gap-4 p-8 w-full bg-white">
                     <div>
@@ -52,7 +34,5 @@ export const UserContact = () => {
                     <Button className="h-10">Guardar datos</Button>
                 </div>
             </div>
-        </LayoutAsideSection>
-
     )
 }
