+++ src/modules/core/components/index.ts
@@ -7,7 +7,7 @@ import { DataTable } from './DataTable/table'
 import * as DataTableComponents from './DataTable/sections'
 import * as DataTableSections from './DataTable/sections'
 import { UserProfile } from './UserProfile/UserProfile'
-import { UserContact } from './UserContact/UserContact';
+import { UserContact } from './UserProfile/UserContact';
 
 
 export { NavBarCustom, UserDropdown, HeaderSection, LoginComponent, CardGroup, DataTableComponents, DataTable, DataTableSections, UserProfile, UserContact }
